Change BlogApp.Core to DDD
