+++ BlogApp/src/BlogApp.Core/Base/Methods/IChangeAsync.cs
@@ -1,0 +1,6 @@
+namespace BlogApp.Core.Base.Methods;
+
+public interface IChangeAsync<TEntity>
+{
+    public Task ChangeAsync(int id, TEntity entity);
+}
