+++ BlogApp/src/BlogApp.Core/Base/Methods/ICreateAsync.cs
@@ -1,0 +1,6 @@
+namespace BlogApp.Core.Base.Methods;
+
+public interface ICreateAsync<TEntity>
+{
+    public Task CreateAsync(TEntity entity);
+}
