+++ BlogApp/src/BlogApp.Core/Base/Methods/IDeleteAsync.cs
@@ -1,0 +1,6 @@
+namespace BlogApp.Core.Base.Methods;
+
+public interface IDeleteAsync<TEntity>
+{
+    public Task DeleteAsync(TEntity entity);
+}
