+++ BlogApp/src/BlogApp.Core/Base/Methods/IGetAllAsync.cs
@@ -1,0 +1,6 @@
+namespace BlogApp.Core.Base.Methods;
+
+public interface IGetAllAsync<TEntity>
+{
+    public Task<IEnumerable<TEntity>> GetAllAsync();
+}
