+++ BlogApp/src/BlogApp.Core/Base/Methods/IGetByIdAsync.cs
@@ -1,0 +1,6 @@
+namespace BlogApp.Core.Base.Methods;
+
+public interface IGetByIdAsync<TEntity>
+{
+    public Task<TEntity> GetByIdAsync(int id);
+}
