--- BlogApp/src/BlogApp.Core/Dtos/LoginDto.cs
@@ -1,17 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace BlogApp.Core.Dtos
-{
-    public class LoginDto
-    {
-        [Required]
-        public string? Name { get; set; }
-        [Required]
-        public string? Email { get; set; }
-
-    }
-}
