+++ BlogApp/src/BlogApp.Core/Dtos/Models/LoginDto.cs
@@ -1,0 +1,12 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace BlogApp.Core.Dtos.Models;
+
+public class LoginDto
+{
+    [Required]
+    public string? Name { get; set; }
+    [Required]
+    public string? Email { get; set; }
+
+}
