--- BlogApp/src/BlogApp.Core/Dtos/RegistrationDto.cs
@@ -1,16 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace BlogApp.Core.Dtos
-{
-    public class RegistrationDto
-    {
-        [Required]
-        public string? Name { get; set; }
-        [Required]
-        public string? Email { get; set; }
-    }
-}
