--- BlogApp/src/BlogApp.Core/Models/Topic.cs
@@ -1,16 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace BlogApp.Core.Models
-{
-    public class Topic
-    {
-        [Key]
-        public int Id { get; set; }
-        [Required]
-        public string? Name { get; set; }
-    }
-}
