--- BlogApp/src/BlogApp.Core/Models/User.cs
@@ -1,15 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace BlogApp.Core.Models;
-
-public class User
-{
-    [Key]
-    public int Id { get; set; }
-    [Required]
-    public string? Name { get; set; }
-    public string? AvatarUrl { get; set; }
-    [Required]
-    public string? Email { get; set; }
-}
