--- BlogApp/src/BlogApp.Core/Repositories/Base/IGetByEmailAsync.cs
@@ -1,6 +1,0 @@
-namespace BlogApp.Core.Repositories.Base;
-
-public interface IGetByEmailAsync<TEntity>
-{
-    public Task<TEntity> GetByEmailAsync(string email);
-}
