--- BlogApp/src/BlogApp.Core/Repositories/ITopicRepository.cs
@@ -1,14 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using BlogApp.Core.Models;
-using BlogApp.Core.Repositories.Base;
-
-namespace BlogApp.Core.Repositories
-{
-    public interface ITopicRepository : IGetByIdAsync<Topic?>
-    {
-        
-    }
-}
