--- BlogApp/src/BlogApp.Core/Services/ITopicService.cs
@@ -1,13 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using BlogApp.Core.Models;
-
-namespace BlogApp.Core.Services
-{
-    public interface ITopicService
-    {
-        public Task<Topic?> IGetByIdAsync(int id);
-    }
-}
