+++ BlogApp/src/BlogApp.Core/Topic/Repositories/Base/ITopicRepository.cs
@@ -1,0 +1,9 @@
+namespace BlogApp.Core.Topic.Repositories.Base;
+
+using BlogApp.Core.Topic.Models;
+using BlogApp.Core.Base.Methods;
+
+public interface ITopicRepository : IGetByIdAsync<Topic?>, IGetAllAsync<Topic?>
+{
+    
+}
