+++ BlogApp/src/BlogApp.Core/Topic/Services/Base/ITopicService.cs
@@ -1,0 +1,9 @@
+namespace BlogApp.Core.Topic.Services.Base;
+
+using BlogApp.Core.Topic.Models;
+
+public interface ITopicService
+{
+    public Task<Topic?> GetByIdAsync(int id);
+    public Task<Topic?> GetAllTopicsAsync();
+}
